oled/requirements.txt
Adafruit-SSD1306
Adafruit-Blinka==8.39.1
adafruit-circuitpython-busdevice==5.2.9
adafruit-circuitpython-connectionmanager==3.1.0
adafruit-circuitpython-framebuf==1.6.5
adafruit-circuitpython-requests==4.0.0
adafruit-circuitpython-ssd1306==2.12.17
adafruit-circuitpython-typing==1.10.3
Adafruit-PlatformDetect==3.63.0
Adafruit-PureIO==1.1.11
cycler==0.12.1
fonttools==4.51.0
kiwisolver==1.4.5
packaging==24.0
pillow==10.3.0
pyftdi==0.55.4
pyparsing==3.1.2
pyserial==3.5
python-dateutil==2.9.0.post0
pyusb==1.2.1
rpi-ws281x==5.0.0
RPi.GPIO==0.7.1
six==1.16.0
sysv-ipc==1.1.0
typing_extensions==4.11.0
